backend/requirements.txt
Flask==3.1.0
Flask_Cors==5.0.0
psycopg2_binary==2.9.10
python-dotenv==1.0.1
